rename.

+++ Base/Astringent.Game20220410.Protocol/IPlayer.cs
@@ -2,7 +2,7 @@
 {
     public interface IPlayer
     {
-        Regulus.Remote.Property<long> Id { get; }
+        Regulus.Remote.Property<int> Id { get; }
 
         Regulus.Remote.Value<bool> SetDirection(Unity.Mathematics.float3 dir);
     }
